cleaned up a bit
